+++ lib/src/themes/info_popup_arrow_theme.dart
@@ -2,6 +2,8 @@ import 'package:flutter/material.dart' show Color, Colors, CustomPainter, Size;
 import 'package:info_popup/src/constants/popup_constants.dart';
 import 'package:info_popup/src/enums/arrow_direction.dart';
 
+import '../enums/arrow_alignment.dart';
+
 /// [InfoPopupArrowTheme] is used to customize the arrow of the popup.
 class InfoPopupArrowTheme {
   /// [InfoPopupArrowTheme] is creates a [InfoPopupArrowTheme] constructor.
@@ -14,6 +16,7 @@ class InfoPopupArrowTheme {
     this.arrowDirection = ArrowDirection.up,
     this.color = Colors.black,
     this.arrowGap = 0,
+    this.arrowAlignment = ArrowAlignment.center,
     this.arrowPainter,
   });
 
@@ -28,6 +31,9 @@ class InfoPopupArrowTheme {
 
   /// The [arrowGap] add a gap between the arrow and the content.
   final double arrowGap;
+
+  /// The [arrowAlignment] of the arrow indicator.
+  final ArrowAlignment arrowAlignment;
 
   /// The [arrowPainter] is used to draw the arrow indicator.
   final CustomPainter? arrowPainter;